+++ CHANGES.md
@@ -1,5 +1,10 @@
 # fmetools changes
 
+## dev
+
+* Improve performance of `fmetools.features.get_attributes_with_prefix()`.
+* Improve documentation and performance of `fmetools.paramparsing.TransformerParameterParser`.
+
 ## 0.11.1
 
 * Improve DEF line parameter parsing.